Backend/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
openpyxl
google-generativeai
python-dotenv